Added a handler for ctrl-C
This is needed because the way that threads exit is NOT graceful on an interrupt. By catching the signal we can get all of the threads to exit almost immediately. If we don't have a signal mechanism, the motor can end up spinning for full seconds after we want it to stop.